FirestoreProductModel.kt 내부 수정

+++ app/src/main/java/com/company/boogie/models/FirestoreProductModel.kt
@@ -69,41 +69,46 @@ class FirestoreProductModel {
     /**
      * Firestore에서 특정 기자재을 삭제합니다.
      *
-     * @param pid 기자재의 이름입니다.
+     * @param productName 기자재의 이름입니다.
      * @param callback 기자재 삭제 결과에 대한 상태 코드(STATUS_CODE)를 반환하는 콜백 함수입니다.
      */
-    fun deleteproduct(pid: String, callback: (Int) -> Unit) {
-        val imageRef = FirebaseStorage.getInstance().reference.child("Product/${pid}")
-        imageRef.metadata
-            .addOnSuccessListener {
-                imageRef.delete()
-                    .addOnSuccessListener {
-                        db.collection("Product").document(pid).delete()
-                            .addOnSuccessListener {
-                                Log.d("FirestoreProductModel", "[${pid}] 기자재 DB 정보 성공적으로 삭제")
-                                callback(StatusCode.SUCCESS)
-                            }
-                            .addOnFailureListener { e ->
-                                Log.w("FirestoreProductModel", "[${pid}] 기자재 DB 정보 삭제 중 에러 발생!!! -> ", e)
-                                callback(StatusCode.FAILURE)
-                            }
-                    } .addOnFailureListener { e ->
-                        Log.w("FirestoreProductModel", "[${pid}] 기자재 이미지 Storage에서 삭제 중 에러 발생!!! -> ", e)
-                        callback(StatusCode.FAILURE)
-                    }
+    fun deleteProduct(productName: String, callback: (Int) -> Unit) {
+        // Product 컬렉션에서 name 필드가 productName과 일치하는 문서 찾기
+        db.collection("Product").whereEqualTo("name", productName).get()
+            .addOnSuccessListener { documents ->
+                if (documents.isEmpty) {
+                    Log.d("FirestoreProductModel", "[$productName] 해당 제품을 찾을 수 없음")
+                    callback(StatusCode.FAILURE)
+                } else {
+                    // 일치하는 첫 번째 문서에 대해서만 처리
+                    val productDoc = documents.documents.first()
+                    val pid = productDoc.id
+                    val imageRef = FirebaseStorage.getInstance().reference.child("Product/${pid}")
+
+                    imageRef.delete()
+                        .addOnSuccessListener {
+                            productDoc.reference.delete()
+                                .addOnSuccessListener {
+                                    Log.d("FirestoreProductModel", "[$productName] 제품 정보 및 이미지 성공적으로 삭제")
+                                    callback(StatusCode.SUCCESS)
+                                }
+                                .addOnFailureListener { e ->
+                                    Log.w("FirestoreProductModel", "[$productName] 제품 DB 정보 삭제 중 에러 발생!!! -> ", e)
+                                    callback(StatusCode.FAILURE)
+                                }
+                        }
+                        .addOnFailureListener { e ->
+                            Log.w("FirestoreProductModel", "[$productName] 제품 이미지 Storage에서 삭제 중 에러 발생!!! -> ", e)
+                            callback(StatusCode.FAILURE)
+                        }
+                }
             }
-            .addOnFailureListener {
-                db.collection("Product").document(pid).delete()
-                    .addOnSuccessListener {
-                        Log.d("FirestoreProductModel", "[${pid}] 기자재 DB 정보 성공적으로 삭제")
-                        callback(StatusCode.SUCCESS)
-                    }
-                    .addOnFailureListener { e ->
-                        Log.w("FirestoreProductModel", "[${pid}] 기자재 DB 정보 삭제 중 에러 발생!!! -> ", e)
-                        callback(StatusCode.FAILURE)
-                    }
+            .addOnFailureListener { e ->
+                Log.w("FirestoreProductModel", "[$productName] 제품 검색 중 에러 발생!!! -> ", e)
+                callback(StatusCode.FAILURE)
             }
     }
+
 
     /**
      * Firestore에서 기자재 정보를 수정합니다.
@@ -114,9 +119,10 @@ class FirestoreProductModel {
      * @param updatedDetail 수정된 기자재의 상세 정보입니다.
      * @param updatedLocation 수정된 기자재의 위치입니다.
      * @param updatedClassificationCode 수정된 기자재의 상태 코드입니다.
+     * @param updatedProductId 수정된 기자재의 번호입니다.
      * @param callback 상품 정보 수정 상태 코드(STATUS_CODE)를 반환하는 콜백 함수입니다.
      */
-    fun updateProduct(productId: String, updatedName: String, updateImage: String?, updatedDetail: String, updatedLocation: String, updatedClassificationCode: Int, callback: (Int) -> Unit) {
+    fun updateProduct(productId: String, updatedName: String, updateImage: String?, updatedDetail: String, updatedLocation: String, updatedClassificationCode: Int, updatedProductId: String, callback: (Int) -> Unit) {
         val productRef = db.collection("Product").document(productId)
         val imageRef = FirebaseStorage.getInstance().reference.child("Product/${productId}")
 
@@ -132,7 +138,8 @@ class FirestoreProductModel {
                                         "img" to uri.toString(),
                                         "detail" to updatedDetail,
                                         "location" to updatedLocation,
-                                        "stateCode" to updatedClassificationCode
+                                        "classificationCode" to updatedClassificationCode,
+                                        "productId" to updatedProductId
                                     )
                                     val updatedDataMap: Map<String, Any> = updatedData
 
@@ -166,7 +173,8 @@ class FirestoreProductModel {
                                     "img" to uri.toString(),
                                     "detail" to updatedDetail,
                                     "location" to updatedLocation,
-                                    "stateCode" to updatedClassificationCode
+                                    "classificationCode" to updatedClassificationCode,
+                                    "productId" to updatedProductId
                                 )
                                 val updatedDataMap: Map<String, Any> = updatedData
 
@@ -194,7 +202,8 @@ class FirestoreProductModel {
                         "img" to (product?.img ?: ""),
                         "detail" to updatedDetail,
                         "location" to updatedLocation,
-                        "stateCode" to updatedClassificationCode
+                        "classificationCode" to updatedClassificationCode,
+                        "productId" to updatedProductId
                     )
                     val updatedDataMap: Map<String, Any> = updatedData
 
@@ -229,28 +238,33 @@ class FirestoreProductModel {
      * @param callback 상태 코드(STATUS_CODE)를 반환하는 콜백 함수입니다.
      */
     fun updateBanStatus(userEmail: String, isBanned: Boolean, callback: (Int) -> Unit) {
-        // User 컬렉션에서 userEmail을 사용하여 문서 찾기
+        // User 컬렉션에서 email 필드를 사용하여 문서 찾기
         db.collection("User")
-            .whereEqualTo("userEmail", userEmail)
+            .whereEqualTo("email", userEmail)
             .get()
             .addOnSuccessListener { documents ->
-                for (document in documents) {
-                    // 해당 문서의 isBanned 필드 업데이트
-                    db.collection("User").document(document.id)
+                if (documents.isEmpty) {
+                    Log.d("FirestoreProductModel", "[$userEmail] 해당 이메일을 가진 사용자 없음")
+                    callback(StatusCode.FAILURE)
+                } else {
+                    // 일치하는 첫 번째 문서에 대해서만 isBanned 필드 업데이트
+                    val userDoc = documents.documents.first()
+                    db.collection("User").document(userDoc.id)
                         .update("isBanned", isBanned)
                         .addOnSuccessListener {
-                            Log.d("FirestoreProductModel", "사용자 밴 여부 성공적으로 변경 완료")
+                            Log.d("FirestoreProductModel", "[$userEmail] 사용자 밴 여부 성공적으로 변경 완료")
                             callback(StatusCode.SUCCESS)
                         }
                         .addOnFailureListener { e ->
-                            Log.w("FirestoreProductModel", "사용자 밴 여부 변경중 오류 발생", e)
+                            Log.w("FirestoreProductModel", "[$userEmail] 사용자 밴 여부 변경중 오류 발생", e)
                             callback(StatusCode.FAILURE)
                         }
                 }
             }
             .addOnFailureListener { exception ->
-                Log.w("FirestoreProductModel", "필드 불러오는 중 에러 발생", exception)
+                Log.w("FirestoreProductModel", "[$userEmail] 사용자 검색 중 에러 발생", exception)
                 callback(StatusCode.FAILURE)
             }
     }
+
 }
